+++ resources/views/admin/tambah_barang.blade.php
@@ -1,4 +1,7 @@
-@extends('admin.index')
+@extends('layouts.admin.index')
+@section('title')
+Tambah Barang
+@endsection
 @section('isi')
 
 <div class="col-md-12">
@@ -33,7 +36,7 @@
 			      <td>{{$b->HargaNet}}</td>
 			      <td>{{$b->HargaJual}}</td>
 			      <td>{{$b->Stok}}</td>
-            <td><a href="delete_barang/{{$b->id}}">Hapus</a></td>
+            <td><a href="delete_barang/{{$b->id}}" onclick="return confirm('Are you sure you want to delete this item?');">Hapus</a></td>
 			    </tr>
 			  </tbody>
 			  @endforeach
